pyproject.toml
[project]
name = "langchain-course"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "black>=26.1.0",
    "debugpy>=1.8.20",
    "google-generativeai>=0.8.6",
    "isort>=7.0.0",
    "langchain>=1.2.7",
    "langchain-core>=1.2.7",
    "langchain-google-genai>=4.2.0",
    "langchain-ollama>=1.0.1",
    "langchain-openai>=1.1.7",
    "langchain-tavily>=0.2.17",
    "langsmith>=0.6.7",
    "python-dotenv>=1.2.1",
]
